+++ kaboomjs/videojuegos/SnakeGame/code/tags.ts
@@ -1,0 +1,14 @@
+export const SnakeTags = {
+    head: "head",
+    body: "body",
+}
+
+export const FoodTags = {
+    food: "food"
+}
+
+
+export const ScenesId = {
+    main: "game",
+    end: "gameOver"
+}
